Add analytics, admin links and login functionality (#4)
* chore: Update Google Analytics credentials and refactor analytics route

* feat: Add lazy loading for images and update analytics page

* feat: Add @tanstack/react-query for data fetching and caching

* refactor: analytics route

* feat: Update caching strategy for analytics data

* feat: Update analytics page with new metrics and fix sorting order

Refactor the AnalyticsPage component to include new metrics for tracking goal conversions, top sources, and countries. Also, update the sorting order for the visitorsByDate function in the config file. Additionally, handle the error in the analytics route and refactor the SingleMetricStatistic component to handle null data. Finally, add lazy loading for images and update the caching strategy for analytics data.

Closes #123

* refactor: Update sorting order for visitorsByDate function in config.ts

* Update navbar and skills section styling, add new links to siteConfig

* feat: Add caret blinking animation and update education section

Refactor the `tailwind.config.ts` file to add a new animation called "caret-blink" for the blinking caret effect. Also, update the `education.tsx` file to add an `id` attribute to the container div of the education section. This change is made to improve accessibility and allow users to navigate directly to the education section using an anchor link.

* feat: Improve hero section styling and add stats component

Refactor the hero section component to improve the styling and layout. Update the font size and alignment of the heading and paragraph elements. Also, add a new component called Stats to display statistics related to the developer's skills and projects.

* Add: plausible analytics

* add: plausible proxy

* ??

* Change font

* change number ticker & hero

* refactor: change the plausible address to the selfhosted one

* fix layout

* fix layout

* shit

* this works now

* shit...i forgot about this

* feat: change the sidebar and dashboard layout

* type shit

* huuuh?

* Add: services

* add prettier for tailwindcss

* fix: deps

* new design

* Remove unnecesary .env file

* Fix: project page

* chore: chane image on about page

* Add: sidebar admin links and login functionality

* Add: analytics & speed insights from vercel

* fix: forgot to use the newly created provider